containers/hailo-runtime/api/inference.py
"""
推理服务模块
提供图像/视频推理功能
"""

import asyncio
import logging
import tempfile
from pathlib import Path
from typing import AsyncGenerator, List, Optional, Union

import grpc
import numpy as np
from fastapi import APIRouter, File, HTTPException, UploadFile, BackgroundTasks
from fastapi.responses import StreamingResponse

from .config import settings
from .models import InferenceRequest, InferenceResponse, StreamInferenceRequest

logger = logging.getLogger(__name__)

# 尝试导入HailoRT，如果失败则使用模拟实现
try:
    import hailo_platform as hpf
    HAILO_AVAILABLE = True
    logger.info("✅ HailoRT Python API 可用")
except ImportError:
    HAILO_AVAILABLE = False
    logger.warning("⚠️ HailoRT Python API 不可用，使用模拟模式")


class HailoInferenceService:
    """Hailo推理服务类"""

    def __init__(self):
        self.device = None
        self.network_group = None
        self.infer_model = None
        self._ready = False
        self._current_model = None

    async def initialize(self):
        """初始化Hailo设备和推理模型"""
        try:
            if HAILO_AVAILABLE:
                # 初始化Hailo设备
                self.device = hpf.Device()

                # 配置网络 (需要根据实际模型调整)
                # 这里应该加载预编译的HEF模型文件
                # model_path = Path(settings.HAILO_MODEL_PATH) / "model.hef"
                # if model_path.exists():
                #     self.network_group = self.device.create_network_group(
                #         hpf.HEF(model_path)
                #     )
                #     self.infer_model = self.network_group.create_infer_model()
                #     self._current_model = str(model_path)
                #     logger.info(f"✅ 加载模型: {model_path}")
                # else:
                #     logger.warning(f"⚠️ 模型文件不存在: {model_path}")

                logger.info("✅ Hailo设备初始化完成")

            self._ready = True
            logger.info("✅ 推理服务初始化完成")

        except Exception as e:
            logger.error(f"❌ 推理服务初始化失败: {e}")
            raise

    async def cleanup(self):
        """清理资源"""
        try:
            if self.infer_model:
                self.infer_model.release()
            if self.network_group:
                self.network_group.release()
            if self.device:
                self.device.release()

            logger.info("✅ 推理服务资源清理完成")
        except Exception as e:
            logger.error(f"❌ 资源清理失败: {e}")

    def is_ready(self) -> bool:
        """检查服务是否就绪"""
        return self._ready

    async def get_device_status(self) -> dict:
        """获取设备状态"""
        if not HAILO_AVAILABLE:
            return {
                "status": "simulation_mode",
                "device_id": "simulated",
                "temperature": 0.0,
                "power_usage": 0.0,
                "memory_usage": 0.0
            }

        try:
            # 获取实际设备状态
            return {
                "status": "ready" if self._ready else "initializing",
                "device_id": getattr(self.device, 'device_id', 'unknown'),
                "temperature": 0.0,  # 需要根据实际API获取
                "power_usage": 0.0,  # 需要根据实际API获取
                "memory_usage": 0.0   # 需要根据实际API获取
            }
        except Exception as e:
            logger.error(f"获取设备状态失败: {e}")
            return {"status": "error", "error": str(e)}

    async def run_inference(self, request: InferenceRequest) -> InferenceResponse:
        """执行推理"""
        if not self._ready:
            raise HTTPException(status_code=503, detail="推理服务未就绪")

        try:
            # 模拟推理过程
            await asyncio.sleep(0.1)  # 模拟推理延迟

            if HAILO_AVAILABLE and self.infer_model:
                # 实际的Hailo推理逻辑
                # 这里需要根据实际的输入数据格式和模型要求来实现
                pass
            else:
                # 模拟推理结果
                results = {
                    "predictions": [
                        {"class": "example", "confidence": 0.95, "bbox": [0, 0, 100, 100]}
                    ],
                    "inference_time": 0.1,
                    "model": self._current_model or "simulated"
                }

            return InferenceResponse(
                success=True,
                results=results,
                processing_time=0.1,
                model_name=self._current_model or "simulated"
            )

        except Exception as e:
            logger.error(f"推理失败: {e}")
            raise HTTPException(status_code=500, detail=f"推理失败: {str(e)}")

    async def stream_inference(
        self,
        request: StreamInferenceRequest
    ) -> AsyncGenerator[InferenceResponse, None]:
        """流式推理"""
        if not self._ready:
            raise HTTPException(status_code=503, detail="推理服务未就绪")

        try:
            # 模拟流式推理
            for i in range(request.max_frames or 10):
                result = await self.run_inference(
                    InferenceRequest(
                        model_name=request.model_name,
                        input_data=f"frame_{i}"
                    )
                )
                yield result
                await asyncio.sleep(0.1)  # 模拟帧间隔

        except Exception as e:
            logger.error(f"流式推理失败: {e}")
            raise


# 全局推理服务实例
inference_service = HailoInferenceService()

# 创建路由器
router = APIRouter()


@router.post("/single", response_model=InferenceResponse)
async def single_inference(request: InferenceRequest):
    """单次推理接口"""
    return await inference_service.run_inference(request)


@router.post("/image")
async def inference_from_image(
    file: UploadFile = File(...),
    model_name: Optional[str] = None,
    confidence_threshold: float = 0.5
):
    """从图像文件进行推理"""

    # 验证文件类型
    if not file.content_type.startswith('image/'):
        raise HTTPException(status_code=400, detail="只支持图像文件")

    try:
        # 读取图像数据
        image_data = await file.read()

        # 保存临时文件 (可选)
        with tempfile.NamedTemporaryFile(delete=False, suffix='.jpg') as temp_file:
            temp_file.write(image_data)
            temp_file.flush()

            # 转换为numpy数组 (需要实际的图像处理逻辑)
            # image = cv2.imread(temp_file.name)
            # image_array = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)

        # 执行推理
        request = InferenceRequest(
            model_name=model_name,
            input_data=image_data,
            confidence_threshold=confidence_threshold
        )

        result = await inference_service.run_inference(request)

        # 清理临时文件
        Path(temp_file.name).unlink(missing_ok=True)

        return result

    except Exception as e:
        logger.error(f"图像推理失败: {e}")
        raise HTTPException(status_code=500, detail=f"图像推理失败: {str(e)}")


@router.post("/stream")
async def stream_inference_endpoint(request: StreamInferenceRequest):
    """流式推理接口"""

    async def generate_stream():
        async for result in inference_service.stream_inference(request):
            yield f"data: {result.json()}\n\n"

    return StreamingResponse(
        generate_stream(),
        media_type="text/event-stream",
        headers={
            "Cache-Control": "no-cache",
            "Connection": "keep-alive",
        }
    )


@router.get("/status")
async def get_inference_status():
    """获取推理服务状态"""
    return {
        "ready": inference_service.is_ready(),
        "device_status": await inference_service.get_device_status(),
        "current_model": inference_service._current_model,
        "hailo_available": HAILO_AVAILABLE
    }


# gRPC服务函数
async def start_grpc_server(port: int):
    """启动gRPC服务器"""
    try:
        # 这里需要实现实际的gRPC服务器
        # 创建gRPC服务器
        server = grpc.aio.server()

        # 添加推理服务 (需要实现proto文件和service类)
        # inference_pb2_grpc.add_InferenceServicer_to_server(
        #     InferenceServiceServicer(), server
        # )

        # 绑定端口
        listen_addr = f'[::]:{port}'
        server.add_insecure_port(listen_addr)

        # 启动服务器
        await server.start()
        logger.info(f"🚀 gRPC服务器启动在端口 {port}")

        # 等待服务器关闭
        await server.wait_for_termination()

    except Exception as e:
        logger.error(f"❌ gRPC服务器启动失败: {e}")
        raise


# gRPC服务实现类 (需要根据proto文件实现)
class InferenceServiceServicer:
    """gRPC推理服务实现"""

    async def Inference(self, request, context):
        """gRPC推理方法"""
        try:
            # 转换gRPC请求到内部格式
            internal_request = InferenceRequest(
                model_name=request.model_name,
                input_data=request.input_data,
                confidence_threshold=request.confidence_threshold
            )

            # 执行推理
            result = await inference_service.run_inference(internal_request)

            # 转换为gRPC响应格式
            # return inference_pb2.InferenceResponse(
            #     success=result.success,
            #     results=result.results,
            #     processing_time=result.processing_time,
            #     model_name=result.model_name
            # )

        except Exception as e:
            logger.error(f"gRPC推理失败: {e}")
            # 返回错误响应
            context.set_code(grpc.StatusCode.INTERNAL)
            context.set_details(str(e))