containers/hailo-runtime/requirements.txt
# Hailo8 Runtime Container Dependencies

# Core Hailo packages (需要根据实际环境调整)
# hailort==4.23.0  # 这需要实际编译或下载对应版本

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# gRPC support
grpcio==1.59.3
grpcio-tools==1.59.3

# Data processing
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0

# HTTP client and utilities
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Async support
asyncio==3.4.3
aiohttp==3.9.1

# Model serving utilities
onnx==1.15.0
onnxruntime==1.16.3

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# System monitoring
psutil==5.9.6
GPUtil==1.4.0

# Configuration management
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (for caching and logging)
redis==5.0.1
sqlite3  # Built-in